requirements.txt
google-generativeai==0.5.4
pillow==10.3.0
streamlit==1.36.0
streamlit-option-menu==0.3.13
